Move enums to correct package
