+++ models/src/main/java/com/vimeo/networking2/params/SlackLanguagePreference.kt
@@ -1,4 +1,6 @@
-package com.vimeo.networking2.enums
+package com.vimeo.networking2.params
+
+import com.vimeo.networking2.enums.StringValue
 
 /**
  * Language preference of the Slack channel being notified.
